requirements.txt
# Core
python-dotenv==1.0.0
click==8.1.7
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# AI Providers
anthropic==0.8.0
google-genai==0.3.0

# Integrations
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
O365==2.0.27
requests==2.31.0
slack-sdk==3.26.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4
rich==13.7.0
tabulate==0.9.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Optional
fastapi==0.108.0
uvicorn==0.25.0
cryptography==41.0.7
